Add unit tests

+++ plugins/tokens/swap/utils.go
@@ -6,9 +6,9 @@ import (
 	"github.com/tendermint/tendermint/crypto/tmhash"
 )
 
-func CalculteRandomHash(randomNumber []byte, timestamp uint64) []byte {
-	randomNumberAndTimestamp := make([]byte, RandomNumberLength + 8)
+func CalculteRandomHash(randomNumber []byte, timestamp int64) []byte {
+	randomNumberAndTimestamp := make([]byte, RandomNumberLength+8)
 	copy(randomNumberAndTimestamp[:RandomNumberLength], randomNumber)
-	binary.BigEndian.PutUint64(randomNumberAndTimestamp[RandomNumberLength:], timestamp)
+	binary.BigEndian.PutUint64(randomNumberAndTimestamp[RandomNumberLength:], uint64(timestamp))
 	return tmhash.Sum(randomNumberAndTimestamp)
 }
